Blog/requirement.txt
alembic==1.4.1
aniso8601==8.0.0
bleach==3.1.1
blinker==1.4
cffi==1.14.0
click==7.1.1
dominate==2.5.1
elasticsearch==7.5.1
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==3.3.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-MySQLdb==0.2.0
Flask-PageDown==0.2.2
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
Flask-WhooshAlchemy==0.56
Flask-WhooshAlchemyPlus==0.7.4
Flask-WTF==0.14.3
gevent==1.4.0
greenlet==0.4.15
gunicorn==20.0.4
itsdangerous==1.1.0
jieba==0.42.1
Jinja2==2.11.1
Mako==1.1.2
Markdown==3.2.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
Pillow==7.0.0
pycparser==2.20
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.14
sqlparse==0.3.1
urllib3==1.25.8
visitor==0.1.3
webencodings==0.5.1
Werkzeug==1.0.0
Whoosh==2.7.4
WTForms==2.2.1
